+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:35499/91a068ec-b1ea-4699-954e-3613e04c634c?"}
+{"mongoUri":"mongodb://127.0.0.1:40553/eb45ab30-5763-40c6-bfd6-b65e0de515b3?"}
