+++ src/infra/criptograph/bcrypt-adapter/bcrypt-adapter.ts
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt'
-import { HashComparer } from '../../data/protocols/criptography/hash-comparer'
-import { Hasher } from '../../data/protocols/criptography/hasher'
+import { HashComparer } from '../../../data/protocols/criptography/hash-comparer'
+import { Hasher } from '../../../data/protocols/criptography/hasher'
 
 export class BcryptAdapter implements Hasher, HashComparer {
   private readonly salt: number
