+++ src/main/factories/signup/signup.ts
@@ -1,7 +1,7 @@
 import { LogControllerDecorator } from '../../decorators/log'
 import { SignUpController } from '../../../presentation/controllers/signup/signup'
 import { DbAddAccount } from '../../../data/useCases/add-account/db-add-account'
-import { BcryptAdapter } from '../../../infra/criptograph/bcrypt-adapter'
+import { BcryptAdapter } from '../../../infra/criptograph/bcrypt-adapter/bcrypt-adapter'
 import { AccountMongoRepository } from '../../../infra/db/mongodb/account-repository/account'
 import { LogMongoRepository } from '../../../infra/db/mongodb/log-repository/log'
 import { Controller } from '../../../presentation/protocols'
